fix: refactoring and renaming
